requirements.txt
amqp
ansicon
argcomplete
asgiref
billiard
boto3
botocore
Brotli
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
crispy-bootstrap5
cron-descriptor
cryptography
cssselect2
Django
django-celery-beat
django-celery-results
django-cleanup
django-colorfield
django-crispy-forms
django-debug-toolbar
django-extensions
django-filter
django-htmx
django-ratelimit
django-resized
django-storages
django-template-partials
django-tenants
django-timezone-field
django-widget-tweaks
fido2
flower
fonttools
gunicorn
greenlet
humanize
idna
jinxed
jmespath
kombu
packaging
pika
pillow
platformdirs
prometheus_client
prompt_toolkit
psycopg
psycopg-binary
pycparser
pydyf
pyphen
python-crontab
python-dateutil
python-dotenv
pytz
qrcode
redis
requests
s3transfer
setuptools
six
sqlparse
tenacity
tenant-schemas-celery
tinycss2
tinyhtml5
tornado
types-python-dateutil
typing_extensions
tzdata
tzlocal
urllib3
userpath
vine
wcwidth
weasyprint
webencodings
whitenoise
xlwt
zope.event
zope.interface
zopfli
